GameOver overlay is implemented but opens instantly
