+++ fort_macarthur/pubspec.yaml
@@ -29,7 +29,7 @@ dependencies:
   auto_size_text: ^3.0.0-nullsafety.0
   panorama: ^0.4.0
   url_launcher: ^6.0.3
-  flame: ^1.0.0-releasecandidate.12
+  flame: ^1.0.0-releasecandidate.13
 
 dev_dependencies:
   flutter_test:
